fix: handle polls with both POLLIN and POLLHUP
The behaviour of poll() seems to differ between Linux and Darwin when the
child process has closed the pipe and poll() returns both POLLIN and POLLHUP
set on the returned events value. On Linux this is a sign that we can read
from the pipe one more time, but on Darwin it seems that POLLIN remains set
after the final read so we need to add extra logic to make sure to stop
polling for POLLIN after a failed read in the presence of POLLHUP.

Also fix minor formatting issues introduced by Devin.