feat: add rust implementation of t3
This implementation is _inspired_ by the C implementation but has a slightly different interface
and makes of threads and mpsc channels instread of multiple (forked) processes.
It still tries to replicate the backoff ordering mechanism of the C implementation
although the motivation behind that is not entirely clear to me.