fix: improve midline-flush tests
Observed failed tests for the aarch64-linux build job on hydra.nixos.org:
- https://hydra.nixos.org/build/292893102/nixlog/1

I'm confident the usleep() statements within midline-flush.c handles ordering
within a single invocation, but the test does demonstrate that it must be
possible to schedule overlapping program invocations, where the first flush
of a subsequent invocation can be processed in advance of the final flush from
a preceding one.

This update adds a numeric identifier to each hello() and goodbye() statement,
which should help to identify exactly which invocation gets out of order, and
also adds a sleep in the main shell loop to make it that much less likely that
this race will occur.